[PP-1562] Put resources (images, CSS etc.) in pay-product-page subdirectory
The build process generates most of the references to the resources and figures
out the new locations with no intervention (but see caveat below). Some
hardcoded references had to be manually updated.

References in stylesheets to hero images and IE8 fonts not updated to reflect
the new locations (hardcoded elsewhere) but we do not use hero images or
support IE8 so this is OK.