PP-3769 - Introduce privacy notice
Adding a page for the privacy notice and adding to the footer.

Also used position: sticky for the sidebar as the page is long and will
make it easy to navigate, there is pretty good browser support for it
these days and in the spririt of progressive enhancement I think it’s ok
if some users wont get it.