update roadmap for Q4
includes some updates to what we've done this quarter and a new section for next quarter. i've deleted july-september 2018 as the information is broadly available on our features page and is in the tech docs, and keeping it there risks pushing the relevant info too far down the page. have veered towards an optimistic list for next quarter as more likely to get feedback on features, and retrospectively deleting things we haven't done doesn't seem to have been a problem so far. 
not sure if Stripe + Apple/Google pay should be mentioned explicitly at this stage. @tillwirth to confirm.