BAU - middleman/ruby problems with compiling JS
For some reason middleman isnt compiling the JS right the obvious
difference is it’s not minifying it but it also is causing it to error.
So something is up

Rather than solve this, I am just dropping in the compiled code, I know
this is not future forward, but we havent changed this JS is a long time
and wont.