BAU — Check in build directory when building release tarball
For some reasons, sometimes when I do a release, the script the
creates a tarball not of the build directory but of the directory
above, which at that point does not seem to contain a build
directory. It then merrily uploads this to GitHub. This is kind of
dangerous, so check we’re actually in the build directory before
creating the tarball.