WIP: fix redirecting security text to cabinet office
Attempting to fix the build issues and respect the security txt
redirect.

Specifying `.txt.html` instructs the static file generated to use the
appropriate `Content-Type`. Removing the other instructions as they
collide the existing `/security` endpoint (a page defined in this repo)
and cause the build to fail.

Note the `page` config seems to only instruct the `middleman serve`
process, not the static files generated -- this isn't validated but the
behaviour suggests it.