PP-11087: fix redirecting security text to cabinet office
Following on from the good work done in
https://github.com/alphagov/pay-product-page/pull/665

Address the build issues and respect the static security txt redirect.

Specifying `.txt.html` instructs the static file generated to use the
appropriate `Content-Type`.

This removes the other instructions as they collide with the existing
`/security` endpoint (a page defined in this repo) and cause the build to fail.

Note the `page` config seems to only instruct the `middleman serve`
process, not the static files generated -- this isn't validated but the
behaviour suggests it.