PP-11338 add build static website to PR
For a recent story, building the website failed but the PR was only blocked when the the post-merge build and release Github action workflow was run.

We should add a GitHub actions check to PR builds to ensure the products pages can be built using bundle exec middleman build. It would also be useful to run with the --verbose flag, otherwise the cause of a failure might not be shown.

- add build static website to PR test