BAU update performance stats 09.12.2024 bis
With the previous commit, we have already update the performance stats for
this week, however we did aggregate services having similar organisation
name, using the official name.

With this change, we are correcting the count of organisations, having
corrected the name of the organisations directly in Toolbox and re-generated
the count.