PP-14025 Update What’s new section on Home page
With this change, we are updating the What’s new section on the home page in
order to make it more purposeful.

In particular we are mentioning that we are planning to start using Pay By
Bank to make payments faster and cheaper.

We are continuing to link to our roadmap.